Refactor app logic and change ui
