+++ Tests/Contoso.Forms.Test/CrashResultsPage.xaml
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="UTF-8"?>
+<ContentPage xmlns="http://xamarin.com/schemas/2014/forms" xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml" x:Class="Contoso.Forms.Test.CrashResultsPage">
+        
+</ContentPage>
