+++ Tests/Contoso.Forms.Test/CrashResultsPage.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+
+using Xamarin.Forms;
+
+namespace Contoso.Forms.Test
+{
+    public partial class CrashResultsPage : ContentPage
+    {
+        public CrashResultsPage()
+        {
+            InitializeComponent();
+        }
+    }
+}
