+++ Tests/Contoso.Forms.Test/ToggleStatesPage.xaml
@@ -1,0 +1,37 @@
+﻿<?xml version="1.0" encoding="UTF-8"?>
+<ContentPage xmlns="http://xamarin.com/schemas/2014/forms" xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml" x:Class="Contoso.Forms.Test.ToggleStatesPage">
+    <TableView Intent="Form">
+        <TableSection>
+            <ViewCell>
+                <Button AutomationId="EnableMobileCenterButton" Text="Enable Mobile Center" Clicked="EnableMobileCenter" />
+            </ViewCell>
+            <ViewCell>
+                <Button AutomationId="EnableCrashesButton" Text="Enable Crashes" Clicked="EnableCrashes" />
+            </ViewCell>
+            <ViewCell>
+                <Button AutomationId="EnableAnalyticsButton" Text="Enable Analytics" Clicked="EnableAnalytics" />
+            </ViewCell>
+                     <ViewCell>
+                <Button AutomationId="DisableMobileCenterButton" Text="Disable Mobile Center" Clicked="DisableMobileCenter" />
+            </ViewCell>
+            <ViewCell>
+                <Button AutomationId="DisableCrashesButton" Text="Disable Crashes" Clicked="DisableCrashes" />
+            </ViewCell>
+            <ViewCell>
+                <Button AutomationId="DisableAnalyticsButton" Text="Disable Analytics" Clicked="DisableAnalytics" />
+            </ViewCell>
+            <ViewCell>
+                <Label AutomationId="MobileCenterEnabledLabel" x:Name="MobileCenterEnabledLabel" VerticalOptions="Center" HorizontalOptions="Center"/>
+            </ViewCell>
+            <ViewCell>
+                <Label AutomationId="CrashesEnabledLabel" x:Name="CrashesEnabledLabel" VerticalOptions="Center" HorizontalOptions="Center"/>
+            </ViewCell>
+            <ViewCell>
+                <Label AutomationId="AnalyticsEnabledLabel" x:Name="AnalyticsEnabledLabel" VerticalOptions="Center" HorizontalOptions="Center"/>
+            </ViewCell>
+			<ViewCell>
+                <Button AutomationId="DismissButton" Text="Dismiss Page" TextColor="Red" Clicked="DismissPage" />
+            </ViewCell>
+	    </TableSection>
+    </TableView>
+</ContentPage>
